refactor(entity-paper): 调整命名
